Organized Files and fixed warnings
