+++ backend/Ticketing/src/main/java/com/example/demo/Main.java
@@ -1,7 +1,5 @@
 package com.example.demo;
 
-import java.util.Date;
-
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
