--- backend/Ticketing/src/main/java/com/example/demo/MongoConfig.java
@@ -1,13 +1,0 @@
-package com.example.demo;
-
-public class MongoConfig {
-	
-    private static final String MONGO_DB_URL = "localhost";
-    private static final String MONGO_DB_NAME = "ticketing_DB";
-    
-    
-    
-    
-    
-    
-}
