+++ backend/Ticketing/src/main/java/com/example/demo/repository/LocalPassengerRepository.java
@@ -1,7 +1,9 @@
-package com.example.demo;
+package com.example.demo.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
+
+import com.example.demo.controller.LocalPassenger;
 
 
 @Repository
